Show new tweets on top
We add a show more posts button on the top of your timeline to emulate
Twitter's show more.

There's a catch: the DOM diffing is smart enough to append new posts
even though we're passing the whole list of posts. So we have to create
a "new-posts" section.

That allows us to append old posts (for infinite scroll) and prepending
new posts. So, we now have two collections in LiveView. On load, all
posts are in the `posts` assigns. But if you're browsing your Timeline,
and you get a new post, we add it to the `new_posts` assigns and show
the message.