Show recent posts after user clicks on notice
What changed?
============

We were accidentally displaying both the notice of new posts and the
posts themselves. In order to fix that, we create yet another list of
posts called `recent_posts`. These posts are the ones that the user will
see.

And it is a collection we keep going since it's possible for a user
to see new posts and then get newer ones, and then see those. So we want
to keep adding to that collection so long as the user is connected.

But we also want to hide the new posts that the user has not seen yet.
So we keep those in the `new_posts` assigns and only transfer them to
the `recent_posts` assign when the user clicks on the notice.

Note that in order to do this, we no longer set `new_posts` as a
temporary assign. We need to keep those posts in memory until we
transfer them to `recent_posts`.