Add username and name to posts
We add a username and name to posts.

This turned out to be a rather large change. We first removed the
`post.username` field and added `username` and `name` to `user`. We had
to create the association between `post` and `user` and update the
codebase to handle that.

We access the user in the LiveView by grabbing the token in session
(thanks to `phx.gen.auth` for setting it in `UserAuth.log_in_user`) and
fetching the user based on a token from the `Accounts` namespace (thanks
again for `phx.gen.auth`).

We also had to update user registration to run `hash_password/1` outside
of `prepare_changes`. There's an issue open in `phx.gen.auth` that
speaks to this. The problem in the issue is that it's not always desired
to hold a transaction when dealing with hashing passwords. Apparently,
it was put in `prepare_changes` originally as a convenience, so that we
don't do that kind of hashing when validating with LiveView. But we can
simply do the hashing when the changeset is valid. That's what we do in
our case, and that's what Jose reports doing as well. In our case, the
change was driven by a desire to have a `post_factory` that can be
easily inserted (`insert(:post)`) with an associated user that has a
hashed password.