Render list of posts in Timeline
What changed?
=============

We render a list of posts in the Timeline.

We start by testing the rendering of a list of posts from the
`TimelineLiveTest`. There we only check that the post id + post body is
present.

We use that to then step in and test the `PostComponent` to ensure we
have the body and date for the post along with the user's name,
username, and avatar_url.

We also decide to push all the formatting of the date directly into the
`DateHelpers`, so we add a function called `format_short` that shows a
short date like "Nov 5".