Users can like posts
What changed?
============

We let users like posts. We test-drive the implementation with a test
from `timeline_live_test`. Several things to note:

- We could have written the test in the `PostComponent` and
  `phx_target=@myself`. We should refactor to that, but this served to
  show how `phx-value-*` works.

- We add PubSub to the Timeline since this is the first time
  encountering the need for it. We should probably handle receiving
  tweets in your timeline first next time. That way we cross the PubSub
  barrier then.

- Because we didn't refactor our implementation of listing tweets and
  getting new tweets (because we didn't check new tweets coming in),
  we're not using temporary assigns yet. So updating the post in memory
  requires us to traverse all posts and replace the one whose id
  matches.

NOTE: next time handle receiving posts in timeline first.

We also haven't refactored the implementation yet. We could refactor to
handle the event in the post component itself. That would cause us to
move the test in there as well. And extract some helpers for the long
CSS selectors.