Refactor: PostComponent handle like event
What changed?
============

We let the PostComponent handle the like event -- changing the
`phx-value-post_id` to a `phx-target=@myself`.

Note that we didn't have to change the test at all. We don't care who
handles the like so long as everything works as expected.

Typically, I'd recommend moving the test wherever we are handling the
event, which now would mean we should move the test to the
`PostComponentTest`. But the test also covers the `handle_info` callback
in the `TimelineLive` so it's covering both the live view and the live
component, so we leave it in `TimelineLiveTest`.