User can do simple reposting
What changed?
=============

We add simple reposting: it's very similar to a like in this
functionality. A user reposting just updates the repost count in the UI.

I am wondering if we would benefit from a further treatment of this or
if it's more about domain modeling than it is about LiveView testing.