Show a post (with replies)
What changed?
============

A user can now visit a post and see it along with its replies.

We also have to handle a user liking/reposting/replying from the posts
show page.

We add that here.

The one thing we do not have working are liking/reposting of the
replies. They're happening behind the scenes because the PageComponent
handles them, but we're not updating the UI for them.