Liking posts shows as liked
What changed?
=============

A user can now like a post and see that they have liked it. In order to
do that we introduce a new table `likes` and a schema `Timeline.Like`.
Now when a user clicks on a like, we create a `like` that has the
`user_id` and `post_id`. We iterate through the post's like user_ids to
know if the current user has liked it before. Seems pretty brute force
approach, but it seems to work.