User avatar navigates to settings
We add the user's avatar and make it navigate to settings.

In doing so, we see how we can click on a link that is not a live_path
or live_redirect and still use the `follow_redirect/3` helper to get an
`{:ok, conn}` back and assert against the text.