Use temporary assigns for posts
We use temporary assigns for posts. That means we have to choose a
strategy of DOM replacement. We choose to append posts since that's how
we can use InfiniteScroll.

The problem comes when we want to see "new posts". Since LiveView no
longer has them in memory, the new posts get appended rather than
prepended. To get around this, we use a `push_redirect` from the
LiveView to refresh the LiveView without doing a full page reload. It
still has to disconnect and reconnect the LiveView, but it's the only
way to accomplish what we want.

From the docs for `push_redirect/2`:

> The current LiveView will be shutdown and a new one will be mounted in
its place, without reloading the whole page. This can also be used to
remount the same LiveView, in case you want to start fresh.

We do want to start fresh, so that's what we use.