Like posts (part 1)
We test the ability to like the post. This is part one in which we use
`phx-value-*` to send the post to update to the TimelineLive. The test
remains in the `timeline_live` though we do include a test in the
`PostComponent` that checks that we render the correct data on the page.