Like Post (part 2)
We set target `@myself`. The test from the live_view perspective remains
the same. We update the post from the component, and then notify the
parent live view via `send(self(), {:post_updated, post})`. We then
let the Timeline handle that information to update the `posts` assign.

We don't test the function directly in the `PostComponentTest` because
it's internal to the process and ideally not something we want to test
directly.

We continue to have the downside that neither this approach nor the
previous one can distribute the message to other Timelines in realtime.
A different user would have to refresh the page in order to see the
updates.