Like Post (part 3)
We now use Timeline.broadcast_post_update(updated_post) to notify all
subscribers that the post is issued. Since that is happening from the
PostComponent, we write a test asserting that the broadcast happens (by
making the test a subscriber of the event).

We can now remove the `send(self(), {:post_updated, post})` message
since the TimelineLive subscribes to the Timeline events.