Like post (part 4)
We push the broadcast via PubSub into the `Timeline.like_post!`
function. That means our `TimelineLiveTest` is acting like an
integration test since the broadcast is required. But we're also able to
remove the test ensuring the broadcast happens -- that test now lives
in the TimelineTest since that's where the broadcasting is happening.

The TimelineLive test is integration(ish) as can be seen by removing the
broadcast from the `like_post!/2` function. If we remove it, the
TimelineLiveTset fails because it's only with the broadcast that the
Timeline can update the data.