Testing hooks
We show how to use render_hook to test hooks. It's important to use the
element/3 helper in order to ensure we have phx-hook attribute attached
to the correct event.

Unfortunately, we can only test so much. So if deep JS testing is
needed, we have to test that via something like Wallaby with a headless
browser.

Finally, we show that we can sometimes inject parameters into our test
to make it a more explicit test. Before passing `per_page` into our
test, we created 12 records, two of which wouldn't show because our
pagination defaults were 10 records. We can take one of two options:
document that 10 is pagination (which could get out of sync with
reality) or we can allow for injecting the per_page number and make it
explicit in our test. A side benefit is that we can now test pagination
with 2 records rather than needing twelve.