Test rendering of current user's avatar
We test the rendering of the current user's avatar. We start by using
the simple but brittle approach of specifying the full HTML markup. We
then use the `has_element?/3` helper to show we can use a CSS selector
to target the src, and then we extract the element to name it avatar so
that we end with this:

```elixir
avatar = element(view, "view[src*=#{user.avatar_url}]")

assert has_element?(avatar)
```